pyproject.toml
[project]
name = "udemy-django-python-advanced"
version = "0.0.1"
requires-python = ">=3.14"

dependencies = [
    "django~=5.0",
    "djangorestframework",
    "drf-spectacular",
    "Pillow",
    "django-environ"
]

[dependency-groups]
dev = [
    "pytest",
    "factory_boy",
    "Faker"
]

lint = [
    "ruff",
    "mypy",
    "django-stubs[compatible-mypy]",
    "django-stubs-ext",
    "djangorestframework-stubs[compatible-mypy]"
]

prod = [
    "gunicorn",
    "whitenoise",
    "psycopg[binary]"
]

[tool.uv]
default-groups = ["dev", "lint"]

[tool.mypy]
# https://mypy.readthedocs.io/en/stable/command_line.html#cmdoption-mypy-exclude
exclude = [
    ".venv",
    "build",
    "manage.py"
]
ignore_errors = false
warn_return_any = true
disallow_untyped_defs = true
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main",
]

[[tool.mypy.overrides]]
module = ["core.migrations", "core.tests.factories"]
ignore_errors = true

[tool.django-stubs]
django_settings_module = "app.settings.local"

[tool.ruff]
line-length = 100
extend-exclude = [
    "build",
    "*.egg-info",
    "migrations"
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
